% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_macdb.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset}
\usage{
load_dataset(storage_path = here::here("working-data", "data"))
}
\arguments{
\item{storage_path}{where the data is living}
}
\value{
something
}
\description{
Load a dataset
}
