% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_macdb.R
\name{load_table}
\alias{load_table}
\title{Load a dataset}
\usage{
load_table(tablename, storage_path = here::here("working-data", "mcdb"))
}
\arguments{
\item{tablename}{table name}

\item{storage_path}{where the data is living}
}
\value{
something
}
\description{
Load a dataset
}
